# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "dtorkon API",
    "description": "FastAPI backend for the dtorkon mini-blog. The service handles admin-only auth, public/admin posts, site profile editing, analytics, SQLite persistence, searchable publishing, and backend-mediated uploads to Yandex Object Storage over the S3 API.",
    "version": "0.1.0-fastapi"
  },
  "servers": [
    {
      "url": "https://{host}",
      "variables": {
        "host": {
          "default": "example.com"
        }
      }
    }
  ],
  "paths": {
...
    "/api/auth/login": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Create an admin session",
        "operationId": "login_api_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/session": {
      "get": {
        "tags": [
          "Auth"
        ],
        "summary": "Get the current admin session",
        "operationId": "get_session_api_auth_session_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "APIKeyCookie": []
          }
        ]
      }
    },
    "/api/auth/logout": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Delete the current admin session",
        "operationId": "logout_api_auth_logout_post",
        "responses": {
          "204": {
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "APIKeyCookie": []
          }
        ]
      }
    },
...
    "/api/admin/uploads/presign": {
      "post": {
        "tags": [
          "Admin Uploads"
        ],
        "summary": "Prepare a backend upload target for Yandex Object Storage",
        "operationId": "presign_upload_api_admin_uploads_presign_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PresignUploadRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PresignUploadResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "413": {
            "description": "Content Too Large",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "415": {
            "description": "Unsupported Media Type",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Content",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "APIKeyCookie": []
          }
        ]
      }
    },
    "/api/admin/uploads/{asset_id}/content": {
      "put": {
        "tags": [
          "Admin Uploads"
        ],
        "summary": "Upload file content through backend to Yandex Object Storage",
        "operationId": "upload_content_api_admin_uploads__asset_id__content_put",
        "security": [
          {
            "APIKeyCookie": []
          }
        ],
        "parameters": [
          {
            "name": "asset_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Asset Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "410": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Gone"
          },
          "415": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unsupported Media Type"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        }
      }
    },
    "/api/admin/uploads/complete": {
      "post": {
        "tags": [
          "Admin Uploads"
        ],
        "summary": "Mark a backend-mediated storage upload as completed",
        "operationId": "complete_upload_api_admin_uploads_complete_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CompleteUploadRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssetRead"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Content",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "APIKeyCookie": []
          }
        ]
      }
    },
    "/api/admin/assets/{asset_id}/transcribe": {
      "post": {
        "tags": [
          "Admin Uploads"
        ],
        "summary": "Transcribe an audio or video asset through Groq",
        "operationId": "transcribe_asset_api_admin_assets__asset_id__transcribe_post",
        "security": [
          {
            "APIKeyCookie": []
          }
        ],
        "parameters": [
          {
            "name": "asset_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Asset Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssetRead"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        }
      }
    },
    "/api/admin/assets/{asset_id}": {
      "delete": {
        "tags": [
          "Admin Uploads"
        ],
        "summary": "Delete an unused asset",
        "operationId": "delete_asset_api_admin_assets__asset_id__delete",
        "security": [
          {
            "APIKeyCookie": []
          }
        ],
        "parameters": [
          {
            "name": "asset_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Asset Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "AssetRead": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "key": {
            "type": "string",
            "title": "Key"
          },
          "url": {
            "type": "string",
            "title": "Url"
          },
          "mimeType": {
            "type": "string",
            "title": "Mimetype"
          },
          "size": {
            "type": "integer",
            "title": "Size"
          },
          "width": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Width"
          },
          "height": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Height"
          },
          "originalName": {
            "type": "string",
            "title": "Originalname"
          },
          "status": {
            "$ref": "#/components/schemas/AssetStatus"
          },
          "transcriptStatus": {
            "$ref": "#/components/schemas/TranscriptStatus"
          },
          "transcriptText": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Transcripttext"
          },
          "transcriptError": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Transcripterror"
          },
          "transcribedAt": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Transcribedat"
          }
        },
        "type": "object",
        "required": [
          "id",
          "key",
          "url",
          "mimeType",
          "size",
          "originalName",
          "status",
          "transcriptStatus"
        ],
        "title": "AssetRead"
      },
      "AssetStatus": {
        "type": "string",
        "enum": [
          "pending",
          "ready",
          "orphaned"
        ],
        "title": "AssetStatus"
      },
...
      "AttachmentKind": {
        "type": "string",
        "enum": [
          "image",
          "audio",
          "video",
          "file"
        ],
        "title": "AttachmentKind"
      },
...
      "CompleteUploadRequest": {
        "properties": {
          "assetId": {
            "type": "string",
            "title": "Assetid"
          },
          "width": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Width"
          },
          "height": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Height"
          }
        },
        "type": "object",
        "required": [
          "assetId"
        ],
        "title": "CompleteUploadRequest"
      },
...
      "ErrorPayload": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "details": {
            "anyOf": [
              {},
              {
                "type": "null"
              }
            ],
            "title": "Details"
          }
        },
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorPayload"
      },
      "ErrorResponse": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorPayload"
          }
        },
        "type": "object",
        "required": [
          "error"
        ],
        "title": "ErrorResponse"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LoginRequest": {
        "properties": {
          "username": {
            "type": "string",
            "title": "Username"
          },
          "password": {
            "type": "string",
            "title": "Password"
          }
        },
        "type": "object",
        "required": [
          "username",
          "password"
        ],
        "title": "LoginRequest"
      },
...
      "PresignUploadRequest": {
        "properties": {
          "originalName": {
            "type": "string",
            "title": "Originalname"
          },
          "mimeType": {
            "type": "string",
            "title": "Mimetype"
          },
          "size": {
            "type": "integer",
            "title": "Size"
          },
          "kind": {
            "$ref": "#/components/schemas/AttachmentKind"
          }
        },
        "type": "object",
        "required": [
          "originalName",
          "mimeType",
          "size",
          "kind"
        ],
        "title": "PresignUploadRequest"
      },
      "PresignUploadResponse": {
        "properties": {
          "assetId": {
            "type": "string",
            "title": "Assetid"
          },
          "objectKey": {
            "type": "string",
            "title": "Objectkey"
          },
          "uploadUrl": {
            "type": "string",
            "title": "Uploadurl"
          },
          "publicUrl": {
            "type": "string",
            "title": "Publicurl"
          },
          "method": {
            "type": "string",
            "title": "Method"
          },
          "requiredHeaders": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Requiredheaders"
          },
          "expiresAt": {
            "type": "string",
            "title": "Expiresat"
          }
        },
        "type": "object",
        "required": [
          "assetId",
          "objectKey",
          "uploadUrl",
          "publicUrl",
          "method",
          "requiredHeaders",
          "expiresAt"
        ],
        "title": "PresignUploadResponse"
      },
...
      "SessionResponse": {
        "properties": {
          "adminDisplayName": {
            "type": "string",
            "title": "Admindisplayname"
          },
          "expiresAt": {
            "type": "string",
            "title": "Expiresat"
          }
        },
        "type": "object",
        "required": [
          "adminDisplayName",
          "expiresAt"
        ],
        "title": "SessionResponse"
      },
...
      "TranscriptStatus": {
        "type": "string",
        "enum": [
          "idle",
          "processing",
          "ready",
          "failed"
        ],
        "title": "TranscriptStatus"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "APIKeyCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "dtorkon_session"
      }
    }
  },
...
}